requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# HTTP Client
httpx==0.25.2
requests==2.31.0

# AI/ML & Vector Databases
openai==1.3.7
anthropic==0.7.8
chromadb==0.4.18
pinecone-client==2.2.4

# Caching & Message Broker
redis==5.0.1
celery==5.3.4

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Data Processing & Analytics
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2

# Privacy & Compliance
opacus==1.4.0
diffprivlib==0.6.3
opa-client==0.1.0

# External Integrations
slack-sdk==3.26.1
microsoft-graph-python==0.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# System & Process Management
psutil==5.9.6
threading2==0.3.1

# Visualization (for analytics)
plotly==5.17.0

# Environment-specific
# For production, consider using specific versions and removing dev dependencies
